backend/skills/n8n_templates/Facebook-scrapper/requirements.txt
# ============================================================
#  requirements.txt — Facebook Group Events Scraper v4.0
#  Установка: pip install -r requirements.txt
# ============================================================

# ── Playwright: асинхронный браузерный движок ─────────────────
playwright==1.49.1

# ── Stealth-обёртка (скрывает признаки автоматизации) ────────
playwright-stealth==1.0.6

# ── Загрузка .env файла ───────────────────────────────────────
python-dotenv==1.0.1

# ── OpenAI-совместимый клиент (text + vision LLM) ────────────
openai==1.57.4

# ── Чтение ~/.genspark_llm.yaml ──────────────────────────────
PyYAML==6.0.2

# ── Supabase (опционально — upsert событий в raw_posts) ──────
#  Если не используется Supabase — можно не устанавливать.
#  Установить вместе с остальными: pip install -r requirements.txt
supabase>=2.0.0
